pyproject.toml
[project]
name = "coding-assistant"
version = "0.1.0"
description = "An intelligent, extensible coding assistant"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "docker>=7.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "ruff>=0.1.0"]

[project.scripts]
assist = "assistant.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/assistant"]
